Fix: removed warnigns

+++ Tools/ResourceEditor/Classes/Qt/SceneData.cpp
@@ -266,11 +266,11 @@ void SceneData::EditScene(const String &scenePathname)
 		Vector<SceneNode*> tempV;
 		tempV.reserve(rootNode->GetChildrenCount());
 
-		for (int ci = 0; ci < rootNode->GetChildrenCount(); ++ci)
+		for (int32 ci = 0; ci < rootNode->GetChildrenCount(); ++ci)
 		{
 			tempV.push_back(rootNode->GetChild(ci));
 		}
-        for (int ci = 0; ci < tempV.size(); ++ci)
+        for (int32 ci = 0; ci < (int32)tempV.size(); ++ci)
         {
             //рут нода это сама сцена в данном случае
             scene->AddNode(tempV[ci]);
